# perf: add kzg.UnsafeToBytes and kzg.UnsafeFromBytes methods